WTA #82: updated how time is calculated and updated some models and mappers

+++ core/common/src/main/java/com/jacob/wakatimeapp/core/common/data/dtos/CoreDTOs.kt
@@ -13,7 +13,7 @@ data class DependencyDTO(
     val percent: Double,
     val seconds: Int,
     val text: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
+    @SerialName("total_seconds") val totalSeconds: Long,
 )
 
 @Serializable
@@ -26,7 +26,7 @@ data class EditorDTO(
     val percent: Double,
     val seconds: Int,
     val text: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
+    @SerialName("total_seconds") val totalSeconds: Long,
 )
 
 @Serializable
@@ -39,7 +39,7 @@ data class LanguageDTO(
     val percent: Double,
     val seconds: Int,
     val text: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
+    @SerialName("total_seconds") val totalSeconds: Long,
 )
 
 @Serializable
@@ -53,7 +53,7 @@ data class MachineDTO(
     val seconds: Int,
     val text: String,
     @SerialName("machine_name_id") val machineNameId: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
+    @SerialName("total_seconds") val totalSeconds: Long,
 )
 
 @Serializable
@@ -66,7 +66,7 @@ data class OperatingSystemDTO(
     val percent: Double,
     val seconds: Int,
     val text: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
+    @SerialName("total_seconds") val totalSeconds: Long,
 )
 
 @Serializable
@@ -80,7 +80,7 @@ data class EntityDTO(
     val seconds: Int,
     val text: String,
     val type: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
+    @SerialName("total_seconds") val totalSeconds: Long,
 )
 
 @Serializable
@@ -93,5 +93,5 @@ data class CategoryDTO(
     val percent: Double,
     val seconds: Int,
     val text: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
+    @SerialName("total_seconds") val totalSeconds: Long,
 )
